requirements.txt
robotframework>=7.3.2
Appium-Python-Client>=5.2.2